Update/correct comment in frida_project.ipynb
I went back and reviewed the discussion involving use of the extend() method to flatten nested tuples. The list comprehension by itself allows us to unpack the nested tuples and create a new tuple with the tour number, painting, and date directly. I’m sorry for confusing the code with the superfluous extend() stuff.